Fix in udpclient print.

+++ udpclient.cpp
@@ -466,7 +466,7 @@ int main(int argc, char *argv[]) {
 	sender.depttime.tv_usec=PktDept.tv_usec;	
 	istart=realcc();
 	if (dflag) {
-	  printf("[%d] sender.depttime.tv_sec = %06ld sender.depttime.tv_usec = %llu \n", (int)di, (int)sender.depttime.tv_sec,sender.depttime.tv_usec);
+	  printf("[%d] sender.depttime.tv_sec = %06ld sender.depttime.tv_usec = %llu \n", (int)di, (int)sender.depttime.tv_sec,(int)sender.depttime.tv_usec);
 	}
 	rc =sendto(sd, &sender,size1, 0,(struct sockaddr *) &remoteServAddr,sizeof(remoteServAddr));//size> app head
 	istop=realcc();
